Refactor: useContactStore 분리
